Improve options (#2)
* close open handle

* improve options

+++ chunk/create_test.go
@@ -9,7 +9,7 @@ func TestCreateChunkFiles(t *testing.T) {
 	cases := []struct {
 		title              string
 		inputFile          string
-		chunkSize          int
+		chunkSize          int64
 		expectedChunkCount int
 	}{
 		{
